Add ModelAware Mixin
Moved Source-based model awareness to its own trait and interface.
